z120_monitor/requirements.txt
pyyaml>=6.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
ib_insync>=0.9.8
